+++ api/package.json
@@ -15,7 +15,6 @@
     "axios": "^0.21.4",
     "cookie-parser": "^1.4.5",
     "dotenv": "^10.0.0",
-    "ejs": "^3.1.6",
     "express": "^4.17.1",
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^6.0.3",
