+++ api/routes/user/google_login.js
@@ -1,7 +1,5 @@
 const express = require('express');
 const router = express.Router();
-const querystring = require('query-string');
-const axios = require('axios');
 const { USER_GOOGLE_ID, USER_GOOGLE_SECRET, GOOGLE_PUBLIC_KEY, JWT_SECRET } = process.env;
 const jwt = require('jsonwebtoken');
 const argon = require('argon2');
@@ -9,22 +7,7 @@ const { User } = require('../../models/models');
 const { normalizeUsers } = require('../../controller/normalize');
 const { htmlReplacer, transportEmail } = require('../../controller/emailUtils');
 const fs = require('fs');
-
-function getGoogleAuthUrl () {
-    const rootUrl = 'https://accounts.google.com/o/oauth2/v2/auth';
-    const options = {
-        redirect_uri: 'http://localhost:3001/auth/google',
-        client_id: `${USER_GOOGLE_ID}`,
-        access_type: 'offline',
-        response_type: 'code',
-        prompt: 'consent',
-        scope: [
-            'https://www.googleapis.com/auth/userinfo.profile',
-            'https://www.googleapis.com/auth/userinfo.email'                       
-        ].join(' ')
-    }
-    return `${rootUrl}?${querystring.stringify(options)}`
-}
+const { getGoogleAuthUrl, getTokens } = require('../../controller/mediaAuth');
 
 router.get('/auth/google/url', async (req, res, next) => {
     try {
@@ -33,24 +16,6 @@ router.get('/auth/google/url', async (req, res, next) => {
         next(error);
     }
 });
-
-const getTokens = async (code, client_id, client_secret, redirect_uri) => {
-    const url = 'https://oauth2.googleapis.com/token';
-    const values = {
-        code,
-        client_id,
-        client_secret,
-        redirect_uri,
-        grant_type: 'authorization_code'
-    };
-    try {
-        const response = await axios.post(url, querystring.stringify(values), { headers: { 'Content-Type': 'application/x-www-form-urlencoded' }});
-        return response.data;
-    } catch (error) {
-        throw error;
-    }
-   
-}
 
 router.get('/auth/google', async (req,res,next) => {
     const code = req.query.code;
@@ -77,7 +42,7 @@ router.get('/auth/google', async (req,res,next) => {
         const [re, obj] = htmlReplacer(oldText, newText);
         const html = await fs.readFileSync(__dirname + path, 'utf8')
             .replace(re, (match)=>obj[match]);
-        await transportEmail(email, html);
+        await transportEmail(email, html, 'Registro exitoso');
         
         const token = await jwt.sign({ sub: userCreated._id }, JWT_SECRET, { expiresIn: "12h"});
         userCreated = normalizeUsers(userCreated);
